refactor(orders): share order discovery across consumers
