+++ cmd/gc/api_state.go
@@ -6,7 +6,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"log"
 	"os"
 	"path/filepath"
@@ -25,6 +24,7 @@ import (
 	"github.com/gastownhall/gascity/internal/fsys"
 	"github.com/gastownhall/gascity/internal/git"
 	"github.com/gastownhall/gascity/internal/mail"
+	"github.com/gastownhall/gascity/internal/orderdiscovery"
 	"github.com/gastownhall/gascity/internal/orders"
 	"github.com/gastownhall/gascity/internal/runtime"
 	"github.com/gastownhall/gascity/internal/session"
@@ -770,15 +770,17 @@ func (cs *controllerState) OrdersAll() []orders.Order {
 	cfg := cs.cfg
 	cs.mu.RUnlock()
 
-	allAA, err := scanAllOrders(cs.cityPath, cfg, io.Discard, "gc api: order scan")
+	allAA, err := orderdiscovery.ScanAll(cs.cityPath, cfg, orderdiscovery.ScanOptions{
+		OnRigScanError: func(_ string, _ error) error {
+			return nil
+		},
+		OnOverrideError: func(err error) error {
+			log.Printf("gc api: applying order overrides for %s: %v", cs.cityPath, err)
+			return nil
+		},
+	})
 	if err != nil {
 		return nil
-	}
-
-	if len(cfg.Orders.Overrides) > 0 {
-		if err := orders.ApplyOverrides(allAA, convertOverrides(cfg.Orders.Overrides)); err != nil {
-			log.Printf("gc api: applying order overrides for %s: %v", cs.cityPath, err)
-		}
 	}
 
 	return allAA
